requirements.txt
Flask==2.3.0
google-cloud-vision==3.5.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.36.0
google-generativeai==0.3.0
python-dotenv==1.0.0
gunicorn==21.2.0